+++ lib/ui/intro_page.dart
@@ -12,20 +12,12 @@ class IntroPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
       color: bgColor,
-      padding: const EdgeInsets.all(32.0),
-      child: OrientationBuilder(
-        builder: (context, orientation) {
-          final isVertical = orientation == Orientation.portrait;
-
-          return SafeArea(
-            child: Column(
-              children: [
-                Expanded(child: image ?? Container(), flex: isVertical ? 5 : 4),
-                Expanded(child: content, flex: isVertical ? 5 : 6),
-              ],
-            ),
-          );
-        },
+      padding: const EdgeInsets.all(16.0),
+      child: Column(
+        children: [
+          Expanded(child: image ?? const SizedBox.shrink()),
+          Expanded(child: content),
+        ],
       ),
     );
   }