+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+* Fix design padding layout on small device
+* Improve layout Expanded flex
+  
 ## 0.0.2
 
 * Published on Pub